Return the value from the file as backup if the database has no entry
This is to support the case of a background update. It is more likely on
android, but basically, if the app is updated automatically in the background,
the UI won't run before syncs happen, and so the syncs will fail. This falls
back to reading the default json natively in that case (which is fine, since
existing installs are all consistent with the native json).

We can remove this after we are sure that all existing installs have opened the
app at least once.